Per-crack resolution gate, a faintness gate, the calibration gate in CI, and the live capture check
Four changes, three of them found by being asked a question about the numbers
rather than by running the tests.

Per-crack resolvability. The resolution floor was applied per width sample, so a
crack sitting just under it kept only the samples that measured wide and came back
measured from a biased subset: a 0.35 mm crack read 0.382 mm where it should have
been refused. Resolvability is a property of the crack and the frame, not of one
noisy sample. Deciding once on the median of every sample moved the sweep's worst
error from 13.2 to 5.6 per cent and its interval coverage from 95.4 to 98.1.

TOO_FAINT. At a 900 mm stand-off two patches of surface texture survived every
filter and were reported as a 2.0 mm and a 1.5 mm crack. They were wide enough to
measure and 20 grey levels darker than their surround, where real cracks run 70 to
190. A relative test cannot catch it, because at that distance the surface's own
residual spread collapses to one grey level and the artefacts read as 20 sigma.

The calibration gate. Every build photographs a card of printed lines whose widths
are written on it and checks each one comes back with that width. It runs in pytest,
in CI where the image build depends on it, and in the service at startup, where a
failure refuses every upload with CALIBRATION_FAILED. Its own tests make it fail on
purpose: an unreadable marker, a card too far away, and a marker size entered 4 and
10 per cent wrong, which is what a printer's fit-to-page does.

The live capture check. POST /api/capture takes one viewfinder frame and returns the
angle, the scale and the finest crack that view could measure. Same detector and same
arithmetic as the survey, so guidance cannot disagree with the report.

108 tests. Deployed to a Graviton4 c8g.large in us-east-1.

+++ products/hairline/src/hairline/targets.py
@@ -110,12 +110,12 @@ def _compact_target(spec: TargetSpec) -> tuple[np.ndarray, dict[str, Any]]:
     pitch = (spec.page_mm[1] - top - 8.0) / max(1, len(spec.widths_mm))
     x_mm = 8.0
     for i, width_mm in enumerate(spec.widths_mm):
-        width_px = max(1, int(round(width_mm * spec.px_per_mm)))
+        width_px = max(1, round(width_mm * spec.px_per_mm))
         actual_mm = width_px * spec.dot_mm
         y_mm = top + i * pitch
-        x0 = int(round(x_mm * spec.px_per_mm))
-        x1 = int(round((x_mm + spec.line_length_mm) * spec.px_per_mm))
-        y0 = int(round(y_mm * spec.px_per_mm))
+        x0 = round(x_mm * spec.px_per_mm)
+        x1 = round((x_mm + spec.line_length_mm) * spec.px_per_mm)
+        y0 = round(y_mm * spec.px_per_mm)
         page[y0 : y0 + width_px, x0:x1] = 0
         _label(page, spec, x_mm + spec.line_length_mm + 4.0, y_mm + 1.5,
                f"{actual_mm:.3f} mm", pt=7)
